add responseMessageDto in the folder message, refactor the code
